backend/services/auth_service.py
"""
EYEQ – Authentication Service

Handles:
  - Password hashing / verification (bcrypt)
  - JWT access token creation / decoding
  - User lookup helpers
"""

from datetime import datetime, timedelta, timezone
from typing import Optional

import bcrypt
from jose import JWTError, jwt
from sqlalchemy.orm import Session

from backend.models.user_model import User, UserRole
from config.settings import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES
from utils.helpers import get_logger

logger = get_logger(__name__)


# ─── Password Helpers ──────────────────────────────────────────

def hash_password(plain: str) -> str:
    """Return a bcrypt hash of the plaintext password."""
    return bcrypt.hashpw(plain.encode("utf-8"), bcrypt.gensalt()).decode("utf-8")


def verify_password(plain: str, hashed: str) -> bool:
    """Check a plaintext password against its bcrypt hash."""
    return bcrypt.checkpw(plain.encode("utf-8"), hashed.encode("utf-8"))


# ─── JWT Helpers ───────────────────────────────────────────────

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """Create a signed JWT token."""
    payload = data.copy()
    expire = datetime.now(timezone.utc) + (
        expires_delta if expires_delta else timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    payload.update({"exp": expire})
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)


def decode_access_token(token: str) -> Optional[dict]:
    """Decode a JWT token; return None on failure."""
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except JWTError as exc:
        logger.warning("JWT decode error: %s", exc)
        return None


# ─── DB Helpers ────────────────────────────────────────────────

def get_user_by_username(db: Session, username: str) -> Optional[User]:
    return db.query(User).filter(User.username == username).first()


def create_user(db: Session, username: str, password: str, role: UserRole = UserRole.user) -> User:
    """Register a new user; raises ValueError if username already taken."""
    if get_user_by_username(db, username):
        raise ValueError(f"Username '{username}' is already taken.")
    user = User(
        username=username,
        password_hash=hash_password(password),
        role=role,
    )
    db.add(user)
    db.commit()
    db.refresh(user)
    logger.info("New user registered: %s (role=%s)", username, role)
    return user


def authenticate_user(db: Session, username: str, password: str) -> Optional[User]:
    """Return the User if credentials are valid, else None."""
    user = get_user_by_username(db, username)
    if user and verify_password(password, user.password_hash):
        return user
    return None
