requirements.txt
# ─────────────────────────────────────────────────────────────
# EYEQ – Python Dependencies
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9      # for OAuth2PasswordRequestForm

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9     # Uncomment only when PostgreSQL is installed
aiosqlite==0.20.0            # async SQLite (optional)

# ── Authentication ────────────────────────────────────────────
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# ── Environment ───────────────────────────────────────────────
python-dotenv==1.0.1

# ── Computer Vision ───────────────────────────────────────────
opencv-python-headless==4.9.0.80
numpy>=2.0.0

# ── AI / Detection ────────────────────────────────────────────
ultralytics==8.2.18          # YOLOv8
torch==2.3.0                 # CPU build; use torch+cu121 for GPU
torchvision==0.18.0

# ── Alerts / Notifications ────────────────────────────────────
twilio==9.1.0                # SMS

# ── Dashboard ─────────────────────────────────────────────────
streamlit==1.35.0
pandas==2.2.2
requests==2.32.2

# ── Utilities ─────────────────────────────────────────────────
Pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.3.0

# ── Dev / Testing ─────────────────────────────────────────────
pytest==8.2.2
httpx==0.27.0                # async test client for FastAPI
pytest-asyncio==0.23.7
